chore: component segregation... I can read my code now. Still need to add types
