+++ pages/anime/[id].vue
@@ -77,88 +77,23 @@
           </div>
         </section>
 
-        <section
-          class="mb-8 rounded-md bg-white p-8 dark:bg-gray-800 dark:text-slate-50"
-        >
-          <h5
-            class="mb-10 text-center text-xl font-semibold md:mb-6"
-            ref="loadReviewsTarget"
-          >
-            {{ animeData?.reviews }} Review(s):
-          </h5>
-
-          <!-- Comment -->
-          <div class="flex flex-wrap p-1 shadow-lg">
-            <div class="w-full shrink-0 grow-0 basis-auto md:w-2/12">
-              <img
-                src="https://cdn.myanimelist.net/images/characters/9/326322.jpg"
-                class="mb-6 w-full rounded-lg shadow-lg dark:shadow-black/20"
-                alt="Avatar"
-              />
-            </div>
-
-            <div class="w-full shrink-0 grow-0 basis-auto md:w-10/12 md:pl-6">
-              <p class="mb-3 font-semibold">Griffith (Definitely not Femto)</p>
-              <p>
-                Pathetic Anime. It's thanks to this stupid anime people have a
-                hard time deciding if Johan Liebert is the greatest anime
-                villian. NO, NEWS FLASH BUCKOS! It's ME.... ME ME ME ME ME!!!
-                Don't make me do to you what I did to Guts
-              </p>
-            </div>
-          </div>
-        </section>
+        <OverviewReviewCard :id="id" :animeData="animeData" />
       </div>
     </NuxtLayout>
   </div>
 </template>
 
 <script setup>
-import { useIntersectionObserver } from "@vueuse/core";
 import { apiUrls } from "@/api";
 
 const { isDataInCache } = useCheckInCache();
 const route = useRoute();
 const id = route.params.id;
 
-const loadReviewsTarget = ref(null);
-const shouldLoadReviews = ref(false);
-
-const reviewsOverviewKey = `anime-${id}-reviews-overview`;
-const { data, error, execute, pending, status } = await useFetch(
-  apiUrls.animePageReviews(id),
-  {
-    key: reviewsOverviewKey,
-    getCachedData(key) {
-      isDataInCache(key);
-    },
-    immediate: false,
-  },
-);
-
-const { stop } = useIntersectionObserver(
-  loadReviewsTarget,
-  async ([{ isIntersecting }], observerElement) => {
-    shouldLoadReviews.value = isIntersecting;
-    if (isIntersecting) {
-      try {
-        const key = `anime-${id}-reviews-overview`;
-        if (isDataInCache(key)) return;
-        await execute();
-        stop();
-      } catch (error) {
-        console.error("Failed fetching reviews");
-      }
-    }
-  },
-);
-
 const animeKey = `anime-${id}`;
 const { data: animeData } = await useFetch(apiUrls.animePage(id), {
   key: animeKey,
-  getCachedData(key) {
-    return isDataInCache(key);
-  },
+  getCachedData: (key) => isDataInCache(key),
 });
 
 if (!animeData.value) {
